requirements.txt
ytmusicapi==1.11.5
pylast==5.3.0
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0

# Optional: For multi-user support (requires Supabase)
# supabase>=2.3.0
